requirements.txt
uvicorn[standard]
SQLAlchemy 
fastapi
passlib[argon2]
email-validator
psycopg2-binary
python-jose
python-multipart
jinja2